HLE: Clean up SVC dispatch mechanism
